[FIX] account_followup: avoid archived users for followup sender
Issue:
If an archived user is set as the Sales person on an invoice or as the followup responsible on a partner, it will be the one sending the automatic followups.

Steps to reproduce:
- Create a partner and an overdue invoice for this partner
- Change the "Salesperson" of the invoice to another user
- Archive this user
- Accouting > Customer > Followup Reports
- Click on the partner created earlier
- Click the actions and "Process Automatic Follow-ups"
- [17.0] Traceback
- [18.0+] The send message is from the user that was archived

Cause:
`_get_followup_responsible()` does not check is the users it returns are active or not.

Solution:
Create an iterable with all the possibilities and iterate on it to return the first active user in the list.
Fallback on `self.env.user`.

opw-5153159

closes odoo/enterprise#100657

X-original-commit: 0d540462ab459f2e92513989a97800a0d01095e4
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>