[FIX] l10n_fr_fec_import: FEC matching numbers should be set to False
When a FEC file imports journal items, it will create them to draft with
reconciliations stated by a matching number starting with "I" (also the case
for other some types of imports). Once the entries are posted, the system will
create odoo reconciliation records for those journal items.

The issue lies in this part of the code removing the matching when performing
reconciliations for the same matching number
https://github.com/odoo/odoo/blob/9d34c7ee85d105a56c3393a3dbb3e5d62d4d634a/addons/account/models/account_partial_reconcile.py#L206

Most of the time it's not a problem for the auto reconcile mechanism of
FEC. However since this mechanism reconciles grouped on matching + account, we
can end up in a case where a first reconciliation on account A will remove all
matching number similar but on account B leading to no automatic reconciliation
for them.

This commit provides a test to reproduce this issue.

opw-5086823

closes odoo/enterprise#99752

X-original-commit: 77cd2295d596d160edb332e799b4efd56e3ce8c8
Related: odoo/odoo#236186
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>