[FIX] spreadsheet_edition: remove ugly css rule
Because of a really strong rule in o_spreadsheet lib that forced the
box-sizing property pretty much everywhere, we came up with a super
dense rule to counteract it inside odoo and specifically inside the side
panels. This commits aims to simplify it at best with the common
denominator of those rules.

Task-4878174

closes odoo/enterprise#99301

X-original-commit: 7e53171bb060d105197b875430cd38d481471148
Related: odoo/odoo#235367
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>