[IMP] account_accoutant: Only use account name for automatic reconcile model creation
When we detect a pattern, we are creating
automatically a new reconcile model, before this commit
the name of this new reconcile model was the display name
of the account (which is including the account code), with
this commit, we are only using the account name.

By doing that, it adds fewer noises to the reconcile model
names.

task-5226782

Part-of: odoo/enterprise#99529
Related: odoo/odoo#235767
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>