[FIX] project_account_budget: fix test case on adding project
In commit [1], a new test was added, but it fails because the result of
`template_values` does not match the expected outcome. The failure occurs
at line [2], where the method `_get_profitability_values` is called with the
project's environment. This environment includes the superuser, which causes
the discrepancy in the result.

This commit resolves the issue by calling the project (self.project_goats) with
the user set to `self.user_projectuser`.

[1]: https://github.com/odoo/enterprise/commit/29c6e1d648218219c718a0448fffb8b8444af5df
[2]: https://github.com/odoo/odoo/blob/fad692777e90f3c91f91d2a91e706f8bdb64a08a/addons/project/models/project_update.py#L116

runbot-233851

closes odoo/enterprise#98897

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>