[FIX] l10n_mx_edi_stock: fix Carta Porte print
Steps to reproduce:
- Install Accounting and l10n_mx_edi_stock
- Switch to a Mexican company (e.g. ESCUELA KEMPER URGATE)

- In Accounting settings, select "Solucion Factible" as PAC
- Create a product:
  * Product Name: XYZ
  * Product Type: Goods
  * UNSPSC Category: 25101929 Mobile office van
  * Tariff Fraction: 0101299999 Los demás.
  * UMT Aduana: Units
  * Weight: 1.00

- Go to "Inventory / Operations / Transfers / Deliveries"
- Create a delivery:
  * Delivery Address: [a Mexican contact] (e.g. INMOBILIARIA CVA)
  * Product: Product XYZ
  * Transport Type: Federal Transport
  * Vehicle Setup: [something] (e.g. [ABC123] DEMOPERMIT)
  * Gross Vehicle Weight: [anything]
  * Distance to Destination (KM): [anything]
- Validate the delivery
- Generate Delivery Guide
=> CFDI document should be created and signed by the government
- Print Carta Porte

Issue:
The "CCP ID" in the PDF is different than the one in the CFDI sent
to the government.
Each time the Carta Porte is printed, a new "CCP ID" is generated
and use in the PDF.
This also leads to a different QR-code in each PDF.

Cause:
When the Carta Porte is printed, instead of fetching the data from
the existing CFDI document, they are recomputed.
The problem is that "l10n_mx_edi_idccp" field is a non-stored computed
field and its value is a RFC 4122 identifier.
Each time the compute method is called, a new identifier is generated.

Solution:
When printing the Carta Porte, retrieve "CCP ID" from the existing
CFDI document.

opw-5208648

closes odoo/enterprise#99141

X-original-commit: 5b92ea8ffe4abe6a0b43de93d2bef48a663a069e
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>