[FIX] {,,l10n_ch_}hr_payroll{,account,}: make "Pay Category" field optional
The field is always required in the employee and contract template form
views, but a Freelancer, or people without occupation should not bother
with it.

There was a check on the pay category during the payslip creation which
is now removed since the payslip is computed based on the pay structure
on the payslip form not the one on the employee profile.

Employees without a pay category will be excluded from pay runs,
therefore the "Pay Category" placeholder in form views has been updated
to indicate so.

Task-5237800

closes odoo/enterprise#98921

Related: odoo/odoo#235991
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>