[FIX] helpdesk_stock_account: add only correct product on credit note
_
## Short functional explanation of the error
Let's say a customer buys 2 products. An SO is created.
This customer wants to refund only one of the 2 products.
He sends a ticket to helpdesk and we click on refund.
Even if we specify the product to refund, this action
creates a credit note containing both products (previously
present on the SO) instead of only the one to refund.

Additionally, when selecting the product, we could see
in the dropdown of suggestions all the existing products,
instead of only the ones related to the SO.

## Reproduction Steps
1. Create a SO containing 2 different products and confirm it.
2. Create a regular invoice and confirm.
3. Go to Helpdesk. Click on the configuration tab, and helpdesk
 teams.
4. Click on your helpdesk team, scroll down. In After-Sales,
check "Refunds".
5. Create a ticket and specify the customer who wants to refund.
Make sure the correct helpdesk team is assigned.
6. Click on refund. It opens the wizard. Specify the product
to refund and the Invoices to Refund.
7. Click on reverse.

### Expected behavior
A credit note containing only the specified product to refund
should be created.

### Unexpected behavior
The created credit note contains both products originally
present on the SO.

## Origin of the issue
When issuing a refund from helpdesk_stock_account, this piece
of code is called:
https://github.com/odoo/enterprise/blob/2051e84c55618c64179c4b9f3e99f4e795bacd32/helpdesk_stock_account/wizard/account_move_reversal.py#L16-L17

which calls the ```reverse_moves``` method in the
helpdesk_account.py file, which itself calls the ```reverse_moves```
method in the account_move_reversal.py file, in the account module,
and so on.
Finally, we arrive in the account_move.py file. In the
```_reverse_moves``` of this file, we can see the code:
https://github.com/odoo/odoo/blob/bee7fc1f955c52a88b527ad9a2ddf0021529bbc7/addons/account/models/account_move.py#L4937-L4947
where we simply copy all the lines of the move in the SO
without filtering them. As a result, we get the lines
of the product we don't want to refund

__
opw-5148789

closes odoo/enterprise#99477

X-original-commit: 57380e312232bee77ca7f54e46263a6c6c51365d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>