[FIX] account_accountant: statement view
This commit: 935b25f
change the view of the bank statement but by doing so they broke the use of the
statement button on the bank rec widget.

This commit will add the old view back and change the view used in the action of
the statement button. (This view has already been added in 19.1 so it's kind of
a backport of https://github.com/odoo/enterprise/commit/2ea690bb5b08ac014f89cc05813c557ccb8832bb)

no task id

closes odoo/enterprise#98980

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>