[FIX] documents: hide action icons in folder selection dialogs
Action icons (like 'Open Folder', 'Rename', etc.) were incorrectly
appearing in list views outside of the main Documents app, for example,
when selecting a folder in a settings menu or a popup dialog.

This was confusing and caused a crash when an icon was clicked,
as the required functionality was not loaded in those contexts.

This fix removes the action icons from these secondary views.
To keep the fix stable-proof, the template of the widget was  modified,
in 19.1 we will fix this in cleaner way.

The icons are now correctly restricted to the main Documents list view,
where they function as intended. Other views (like folder pickers) now
behave as standard selection lists without errors.

Task-5166843

closes odoo/enterprise#97861

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>