[FIX] pos_l10n_se: unknown message type with some blackbox
The serial protocol used with the Swedish blackbox has 2 versions, with
v2 adding some more commands. Before this commit, we assumed that the
blackbox was compatible with v2, causing an 'unknown message type' error
if it only supported v1.

After this commit, we check the protocol version of the blackbox when we
initialise the driver, so that we only send compatible commands when we
register a receipt.

opw-5077448

closes odoo/enterprise#100075

X-original-commit: 1a7d1e83b8a9b33f96884520cf418411645032cf
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>