[IMP] pos_blackbox_be: remove look for update in opening control popup
When introducing the blackbox queue service, we added a check at the
opening of the POS to ensure that the blackbox drivers were up to date.
This check is done at the start of session opening.
It is now not necessary anymore, if the drivers are not up to date,
an error will also be shown to the user when they try to use a blackbox
functionality.

closes odoo/enterprise#99283

X-original-commit: 149de9159ef97593f290021a6e32a6d603485bb4
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>