[FIX] sale_amazon: pass the carrier code in the order fulfillment data
The `POST_ORDER_FULFILLMENT_DATA` feed that is used to push order
delivery info to Amazon should follow the `OrderFulfillment` schema (see
https://images-na.ssl-images-amazon.com/images/G/01/rainier/help/xsd/release_4_1/OrderFulfillment.xsd),
but it was missing the `CarrierCode` element, which is required in some
countries.

This commit adds the missing element to the payload, with the formatted
carrier name as a value. If the carrier name cannot be matched, "Other"
is used as a fallback to signal Amazon that they should rely on the
`CarrierName` instead.

closes odoo/enterprise#99140

X-original-commit: 854c1b4b4c91dd54971d64545919f433ba2b760a
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>