[FIX] hr_contract_salary_payroll: Avoid `savepoint` in loops.
Savepoints are used in multiple computes related to salary simulation
for an offer, because it creates and updates existing records, but
those should not be committed to the database. Those are currently made
for each offer that needs to be recomputed.
Frequent savepoint, without commit/rollback of the whole transaction,
are a performance issue at the database level.

To partially alleviate this issue, this commit creates 1
savepoint for the whole compute, instead of per record to compute.

closes odoo/enterprise#99080

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>