[FIX] knowledge: avoid race condition highlighted by tour
Before this commit, the `test_knowledge_main_flow` tour sometimes
failed. That tour creates articles, then d&d one of them in the
sidepanel to re-organize articles, and then clicks on a previously
created article to continue editing it. The tour failed on that
step, as the previously created clicked article wasn't properly
selected/displayed in the editor.

The race condition was that two (non synchronized) calls to the
`load` function of the model were done: one after the move because
the move could have altered the displayed article (1) and one
because we selected another article to open by clicking in the
sidebar (2). (1) is done without resId (reload the current article)
and (2) is done with the id of the clicked article.

Depending on the order these two calls are done, we end up with
the clicked article displayed (if (2) is done after (1)), or with
the current article still displayed, but reloaded ((1) done after
(2)).

This commit fixes the race condition by forcing the reload of the
current record for (1), instead of blindly reloading the model,
which might have changed/been requested something else meanwhile.

runbot error~182073

closes odoo/enterprise#99335

X-original-commit: fe15c7887d99441310f995d90af29fc94647027e
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>