[FIX] account_bacs: round amount to avoid float precision error
Issue description:
When creating a BACS batch payment that contains a payment with an amount
that can't be represented well in float (like 645.30), the generated BACS
batch file will have a wrong amount (due to float precision) as the
amount is represented in pence.

Steps to reproduce:
1. Create a BACS vendor payment with amount = 645.30
2. Add this payment to a BACS batch payment.
3. Confirm the batch to generate the export file.
In the file you will notice that the amount in the payment line is 64529
pence instead of 64530.

opw-5159413

closes odoo/enterprise#99333

X-original-commit: 515fcb80b84edf014de18c46d9036550c220c54b
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>