[FIX] helpdesk_timesheet: fix validation Error for analytic account
When creating an analytic line from a helpdesk ticket,
we assigned the account_id from the project's account_id
during the upgrade. However, in the standard code,
the account_id is later overridden and updated from
ticket.analytic_account_id, which is null.
As a result, the constraint
"At least one analytic account must be set" is triggered.
see:
https://github.com/odoo/enterprise/blob/dcfef2cc462631f376a596a7c85ae483826835ad/helpdesk_timesheet/models/account_analytic_line.py#L119

Multiple upgrade request failed due to this.

closes odoo/enterprise#99209

X-original-commit: 1f4b32e8f5a27b23c9d5fd2eaf108669ff364786
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>