[FIX] hr_payroll_attendance: make sure ruleset_id has right group
Before this commit, the `ruleset_id` field was only accessible to hr
manager and not payroll user with hr user right access which leads an
inconsistency with the other fields in `hr.version` and `hr.employee`
models.

This commit alters the groups of `ruleset_id` field in `hr.employee` and
`hr.version` models to set payroll user instead of hr manager once
`hr_payroll` and `hr_attendance` modules are installed. That is, once
`hr_payroll_attendance` is installed.

closes odoo/enterprise#99086

Related: odoo/odoo#234932
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>