[FIX] pos_blackbox_be: handle offline in syncAllOrders
When trying to sync orders while being offline, a `ConnectionLostError`
is raised. This error was not handled in the pos_blackbox_be module.
Now, if an order was not signed correctly by the blackbox, we put its
state back to "draft", allowing the cashier to retry later
(when the connection to bbox is re-established).

closes odoo/enterprise#99202

X-original-commit: c298970
Related: odoo/odoo#235147
Signed-off-by: David Monnom (moda) <moda@odoo.com>