[FIX] pos_blackbox_be: deleted order callback
Before this commit, when trying to send a pro forma for an order that
had been deleted while the pro forma call was in the queue, a JS error
would occur because the callback of the call would try to access the
order which was no longer existing in the frontend.

This is now fixed by checking that the order is still present before
accessing it in the callback.

closes odoo/enterprise#99302

X-original-commit: a518a5b8a0c7ecd6fa47053b126c9a57a7777a05
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>