[FIX] stock_barcode_picking_batch: await processBarcode
Commit bc9247d46225c696842bc7b0e3c883320231ab1b has introduced an
override of the `processBarcode` method. However, it does not
return nor await the super call. In particular, in teh case where the
super call should be done the overrides returns "undefine" rather than a
promess to await and hence that call is not awaited anymore.

Note: This error has been noticed from the fact that the test
`test_barcode_batch_scan_lots` sometimes fails on step 29/31.

runbot-233631

closes odoo/enterprise#99451

X-original-commit: b5d49b900b6e31acb87a33a2943e281419355ceb
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>