[REV] account_iso20022: Wrong XML generated for Switzerland
This reverts commit 753758a93b63e490a19ab11a9d463693e3a30817.

This commit was introduced after this fix c160b2e. Though that original fix was correct, it had some unwanted side-effects: due to the payment method being forced on some payments in order to use SEPA, the sepa_pain_version field was used to generate the corresponding XML nodes, most of the time keeping its default value of pain.001.001.09. For Swiss banks supporting pain.001.001.09 (which becomes mandatory in November 2026), it was not a problem, and everything worked fine. This was the case of the customer for whom the fix was made (ticket 4535542). For the ones still not supporting it, and expecting pain.001.001.03, the bank refused the file, since the it contained unsupported nodes, like BICFI, or a subnode to ReqdExctnDt.

The commit we revert here tried to patch the symptoms without really understanding the cause of the issue, by not forcing the payment method everywhere. It breaks again the case of the original ticket (because ScvLvl is not passed to "SEPA" on EUR payments), and essentially makes no sense. We revert it in favor of a better fix.

ticket-4535542

Part-of: odoo/enterprise#99150
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>