[FIX] account_iso20022: recompute missing end-to-end IDs for legacy payments
The migration script [1] for the end-to-end feature creates the
`end_to_end_uuid` field even when `account_online_payment` is not
installed. In such databases, the field is added but remains NULL for
existing payments.

If these legacy payments are later added to a batch, a traceback occurs
because the batch logic assumes that `end_to_end_uuid` is always set [2].

This commit generates a UUID on the fly when exporting a payment whose
`end_to_end_uuid` is empty, ensuring that batch creation works correctly
even when `account_online_payment` was not installed.

[1]: https://github.com/odoo/upgrade/blob/ed83d47b2aaa63113ea86ea83d1ca2c886f60063/migrations/account_iso20022/saas~18.5.1.0/pre-migrate.py#L10
[2]: https://github.com/odoo/enterprise/blob/ebdc2fe2ade72b1505dac5490b7df2a1d06fcb4b/account_iso20022/models/account_journal.py#L289

no task-id

closes odoo/enterprise#99364

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>