[FIX] sign: detect typo by running tour as Laurie
By running the test_sign_flow tour as a regular user, a typo is detected
by the tests in the sign.send.request wizard.
The issue is that, on the model `res.partner`, the field `user_id` is
the Salesperson associated to this Contact, whereas the field `user_ids`
contain the `res.users` that inherit this Contact, and `main_user_id` is
the most appropriate User of `user_ids` when we need only one.

When testing as admin, all fields are the admin user, whereas when
testing as Laurie Poiret (or any regular user), the Salesperson is still
the admin while `main_user_id` is Laurie Poiret.

A regular user can access only its own `sign_signature` field, while
an administrator can access the `sign_signature` of all users, so this
bug remained undetected:

On a runbot, log in as `admin` to change the sales person of Marc Demo
to Mitchel Admin, then log in as `demo` and click on "Sign Now" on the
`Rental_Agreement.pdf` template.

    You do not have enough rights to access the field "sign_signature"
    on User (res.users). Please contact your system administrator.

    Operation: read
    User: 5
    Groups: allowed for groups 'Role / Administrator'

closes odoo/enterprise#99408

Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>