[FIX] voip: avoid crash on incoming calls for regular users
In order to prevent users from tampering with their voip.call records,
they aren't given direct write access to them.

This leads to crashes in the get_contact_info function, which attempts
to update the partner_id field of call records.

This commit resolves the issue by switching to sudo mode for setting the
partner_id.

closes odoo/enterprise#99429

Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>