[FIX] hr_recruitment_reports: error when copying recruitment stage report record
When the user tries to perform create or copy operations on
the "hr.recruitment.stage.report" model, a traceback appears.

Steps to reproduce the error:
- Install "hr_recruitment_reports" module with demo data
- Go to Recruitment > Reporting > Hiring Velocity > switch to cohort view >
  click on any record > select any record > duplicate

Traceback:
"ObjectNotInPrerequisiteState
cannot insert into view 'hr_recruitment_stage_report'"

"hr.recruitment.stage.report" model is "_auto=False",
meaning that no database table is created for this model.
When the user attempts to create or duplicate a record of that model,
It will lead to the above traceback.

sentry-7024266324

closes odoo/enterprise#99499

Signed-off-by: Arthur Pierrot (arpi) <arpi@odoo.com>