[FIX] account_batch_payment: wrong sequence number in vendor batch
Currently when creating a vendor batch payment, the system may
incorrectly use the inbound batch payment sequence to generate the
batch name

Steps to reproduce:
- Create a Vendor payment.
- Create a Vendor batch payment, add the payment, validate it.

Issue: Name has been set using the incoming batch payment sequence

This occurs because `batch_type` is set to readonly when a payment is
added to the list. As result, the current batch type is never sent to the
backend that will use the default 'inbound'.

opw-5128426

closes odoo/enterprise#99751

X-original-commit: 867a5f0ff1edb2592579e14545691a9b4094bec3
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>