[FIX] l10n_de_reports: datev export when company name contains dot
Currently companies with a dot ('.') in the name cannot export the
DATEV report due to a traceback.

Steps to reproduce:
- Select a German (DE) Company having a dot in the name.
- Open Accounting / Reporting / Ledgers / General Ledger
- Click cog > Datev DATA (zip)

Issue:

A Traceback will occur
`Error "ValueError: too many values to unpack (expected 2)"`

This occurs because the system separate the filename from the extension
without halting to the first `.`, leading to the error.

opw-5232073

closes odoo/enterprise#99578

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>