[FIX] l10n_in_reports: fix incorrect domain in tax return checks
This commit fixes traceback while opening the tax return for indian
company and also fixes incorrect model name for uqc check.
Traceback
  File /usr/lib/python3.12/json/encoder.py, line 180, in default
    raise TypeError(f'Object of type {o.__class__.__name__} '
TypeError: Object of type DomainAnd is not JSON serializable

This Traceback is due to the use of `Domain.And` in the check of uqc code
from commit :-
https://github.com/odoo/enterprise/commit/5069c133def1ab9f3288fe42c9c97d192d263321

Reason for traceback
The checks uses `_get_records_action` to open the records which internally adds
the `dict(self.env.context)` which does incorrect json serialization.

Also this commit replaces `uom.uom` model inplace of `account.move.line` to
open uom instead of lines.

closes odoo/enterprise#99628

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>