[FIX] pos_blackbox_be: wrong error message on timeout
Before this commit, in the case when there is a timeout communicating
with the Blackbox in the POS, an incorrect error message was shown
stating that "The IoT Box is connected but the Fiscal Data Module
isn't". This message should only be shown when we receive a reply from
the IoT box, but it tells us it cannot find the Blackbox.

closes odoo/enterprise#99687

X-original-commit: 09153fb1cd82f6a98f8e0f0558fb3ab2debfc119
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>