[FIX] web_studio: reportEditor focus iframe's doc on insert
On a new report, add a X2Many table in a new Report.
In many cases there will be some issues with the selection as, when inserting the Element via the command of the report Editor
we explicitly focus the editable of the html_editor.

We need the document inside the iframe to get the focus, because our flow implied to click on some popover bound to the
main window. But focusing the editable element changed the selection.

So, instead, we focus the iframe's inner window, and the selection stays at the right place, and the user can
immediately interact with it (by continuing typing after the insertion)

task-5159482

closes odoo/enterprise#99430

X-original-commit: 443a68c
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>