[FIX] sign: ensure invalid PDF tests works across different PyPDF versions
Case 1:
When pypdf2_2 is not installed, it falls back to using the pypdf package. (Ref1)
This triggers a warning during the test case when it tries to parse an invalid PDF.

This commit suppresses the warning to ensure the test runs without warnings.

Ref1: https://github.com/odoo/odoo/blob/26a5384af0af8fc6e6b5a10bea277f937e2b3481/odoo/tools/pdf/__init__.py#L42-L46

Case 2:
With PyPDF2===1.26.0, the line at [1] raises a `PyPDF2.utils.PdfReadError: EOF marker not found`.
This exception is not handled by the same except block but is instead
handled later in the flow at [2].
As a result, the test raises a UserError, causing the assertion to fail.

On the other hand, when pypdf2_2 or pypdf is installed, the PdfFileReader
raises a UnicodeDecodeError, which is then handled as a ValidationError,
allowing the test to pass as expected.

This commit adds handling for PdfReadError in method get_valid_pdf_data() to
unify the behavior across all supported PyPDF versions.

1: https://github.com/odoo/enterprise/blob/f79601c62ca629dc01a5c1ad5520b0bb44a169d0/sign/utils/pdf_handling.py#L27
2: https://github.com/odoo/enterprise/blob/f79601c62ca629dc01a5c1ad5520b0bb44a169d0/sign/models/sign_document.py#L433-L437

Runbot-234021, 234022

closes odoo/enterprise#99757

Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>