[IMP] l10n_be_reports: Remove exempted taxes from the partner vat listing report
Currently, the behavior is wrong and 0% taxes appear in the report (if the cumulative base for a partner is > 250)

A distinction needs to be done here. The operations are taxable (even at a zero rate) or exempt.

We shall rely on the Tax Category Code (E = Exempt)

When a tax belongs to E, it cannot appear in the Partner VAT Listing. It is exempt from taxation. It does not open the right to deduction of vat on purchases.

task-5269970

closes odoo/enterprise#99803

Signed-off-by: John Laterre (jol) <jol@odoo.com>